// interfaces/IShift.ts
export interface IShift {
  name: string;
  lastName: string;
  email: string;
  dateAndTime: Date;
  numberOne: string;
  numberTwo: string;
  numberThree: string;
  numberFour: string;
  numberFive: string;
  numberSix: string;
}

export interface IShiftWithCode {
  name: string;
  lastName: string;
  email: string;
  dateAndTime: Date;
  code: string;
}